+++ packages/prettier/src/options/sort-miniapp.ts
@@ -4,8 +4,8 @@ import organizeAttributes from '../plugin-organize-attributes'
 // https://prettier.io/docs/options
 // https://github.com/NiklasPor/prettier-plugin-organize-attributes
 export default {
-  parser: 'html',
   plugins: [organizeAttributes],
+  parser: 'html',
   attributeSort: 'ASC',
   attributeIgnoreCase: true,
   attributeGroups: [